Cleanup + new TODOs following peer review
